Call reset on new plan, not new goal
The documentation on `reset()` says:

> called at the beginning of every new navigation, i.e. when we get a
> new global plan

The `reset()` functions were called in `setGoalPose()`, i.e. when the
user gives a new goal. However, when the local planner fails and the
global planner replans, `setGoalPose()` is not called, but `setPlan()`
is. This commit makes sure that `reset()` is also called in this case.